# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Approver",
        "description": "Approver Management",
        "version": "1.0.0"
    },
    "paths": {
        "/api/expense/{id}/approve": {
            "patch": {
                "tags": [
                    "Expense Form"
                ],
                "summary": "Approve Expense",
                "description": "Approve an expense by a specific approver",
                "operationId": "approveExpense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "description": "ID of the approver",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "expense_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "approver_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-10T15:53:46.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-10T15:56:17.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/approval-stages": {
            "post": {
                "tags": [
                    "Approval Stage Form"
                ],
                "summary": "Create a new approver stage",
                "description": "Add approver stage which uses ascending sort by ID",
                "operationId": "addApproverStage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "description": "ID of the approver",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Result of the operation",
                                            "properties": {
                                                "approver_id": {
                                                    "description": "ID of the approver",
                                                    "type": "integer"
                                                },
                                                "updated_at": {
                                                    "description": "Timestamp when the record was last updated",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "created_at": {
                                                    "description": "Timestamp when the record was created",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "id": {
                                                    "description": "Unique identifier of the approver stage",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/approval-stages/{id}": {
            "put": {
                "tags": [
                    "Approver Form"
                ],
                "summary": "Update an approver stage",
                "description": "Updates an approver stage using the specified ID",
                "operationId": "updateApproverStage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the approver stage to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "description": "ID of the approver",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Result of the update operation",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}